meson: Drop gcc 8 support
The libcamera CI has retired gcc 8 testing with the end of life of
Debian Buster at the end of June 2024. As gcc 8 isn't tested anymore, we
can't guarantee it will keep building libcamera correctly. Drop its
support. If anyone still has a compeling use case for gcc 8 support for
libcamera, we will investigate how to restore it in CI.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>