libtuning: Modify ctt_awb.awb() so that it can run in our context
Modify the awb function that was copied from the Raspberry Pi tuning
scripts in a way that it can easily be called from the libtuning code.
In essence the logging was replaced by calls to a python logger and the
need for the Cam object was removed by providing a list of images to the
function.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>