libcamera: Move IPA headers to the libcamera_private dependency
The IPA headers are listed in the libcamera_public and libcamera_private
dependency objects, with the generated headers part of the private
dependency object and the non-generated headers part of the public
dependency object. As neither set of IPA headers are part of the public
API, list them both in the libcamera_private dependency object.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>