libcamera: controls: Add enum names and values map to ControlId
Add to ControlId information about the names and values of enum, in the
event that the ControlId is an enum type. This allows applications to
query the ControlId for the names of the enum values, so that they can
be displayed on a UI, for example. Without this, it was necessary to use
macros of NameOfControlNameValueMap, which is difficult to use and is
very inflexible.

The main map is name -> value, as this was necessary for mapping names
from tuning files to enum values. We reuse this generated code to
reduce complexity, and generate a reverse map from it.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>