libcamera: yaml-parser: Differentiate between empty and empty string
When accessing a nonexistent key on a dict the YamlObject returns an
empty element. This element can happily be casted to a string which is
unexpected. For example the following statement:
yamlDict["nonexistant"].get<string>("default") is expected to return
"default" but actually returns "". Fix this by introducing a empty type
to distinguish between an empty YamlObject and a YamlObject of type
value containing an empty string. For completeness add an isEmpty()
function to be able to test for that type.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>