libcamera: yaml-parser: Differentiate between empty and empty string
When accessing a nonexistent key on a dict the YamlObject returns an
empty element. This element can happily be casted to a string which is
unexpected. For example the following statement:

yamlDict["nonexistant"].get<string>("default")

is expected to return "default" but actually returns "". Fix this by
introducing a empty type to distinguish between an empty YamlObject and
a YamlObject of type value containing an empty string. For completeness
add an isEmpty() function and an explicit cast to bool to be able to
test for that type.

The tests are adapted accordingly. Additional tests for empty entries in
lists and dicts are added to ensure that these still get parsed as empty
strings. This is needed for algorithms without parameters to be parsed
correctly from tuning files.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>