libcamera: converter: Add dw100 vertex map class
Using a custom vertex map the dw100 dewarper is capable of doing
complex and useful transformations on the image data. This class
implements a pipeline featuring:
- Arbitrary ScalerCrop
- Full transform support (Flip, 90deg rotations)
- Arbitrary move, scale, rotate

ScalerCrop and Transform is implemented to provide a interface that is
standardized libcamera wide. The rest is implemented on top for more
flexible dw100 specific features.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>

---

Changes in v2:
- Replaced manual transforms with an affine transformation matrix
- Changed rotation direction to be in sync with the rotation in
  CameraConfiguration::orientation
- Changed offset parameter to be in ScalerCrop coordinates. This is
  easier to explain and has the added benefit, that Scale/Rotate is
always centered to the visible image.
- Improved code comments
- Make dw100VerticesForLength a local function
- Dropped unnecessary includes
- Added documentation

Changes in v0.9
- Include header in meson.build
- Fix black line at top and left when rotation 180 degrees

Changes in v0.8
- Cleanup & formatting

Changes in v0.5
- Fix crash in std::clamp() due to rounding errors