libcamera: geometry: Rename Rectangle's top-left corner
The libcamera::Rectangle class should allow defining rectangles
regardless of the reference system where an instance is used in.

However the class documentation assumes the reference system origin
point to be in the visual top-left side and that rectangles are
constructed by specifying the top-left corner position and sizes.

However this is ambiguous when using a different reference system as,
in example, constructing a rectangle (x, y, w, h) in the following
scenario, with the origin corner in the top-left position, would require
a negative decrement along the y-axis.

         ^
         |
        y|      o------------------
         |      |                 | h
         |      |                 |
         |      -------------------
         |               w
          ------------------------------->
         (0,0)  x

To remove ambiguities, drop references to the visual top-left corner and
define the rectangle's origin corner as the corner closer to the
reference system origin point, and specify the horizontal and vertical
dimensions always increment along their respective axis, as in the
following example.

         ^
         |
         |      -------------------
         |      ^                 | h
         |      |                 |
        y|      o---->-------------
         |               w
          ------------------------------->
         (0,0)  x

         (0,0)  x
           ------------------------------>
          |              w
         y|     o---->-------------
          |     |                 | h
          |     v                 |
          |     -------------------
          |
          V

Also rename for consistency the Rectangle::topLeft() function to
Rectangle::origin() and update its in-tree users.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>