config: Look up logging levels in the configuration file
In addition to LIBCAMERA_LOG_LEVELS environment variable, expose the
same configuration in the configuration file.

Because logging asks for configuration and configuration loading uses
logging, extra care is needed.  We must prevent calling any log function
before configuration is loaded.  Otherwise the log function would ask
for configuration, it would load the configuration files and would try
to log the results, which would cause mutex deadlock due to calling
logging inside logging.  Not logging in configuration is not an option,
at least without introducing complicated mechanism such as delayed
logging, because we need to log at least errors when reading the
configuration.

We must also prevent logging in configuration loading before the
configuration is loaded and available (or there is an error in loading
the configuration).  Otherwise logging would see an empty configuration
and wouldn't be configured as asked.

Preventing logging before configuration is loaded is ensured by asking
for configuration in CameraManager constructor and in IPAManager
constructor (which is mysteriously loaded from CameraManager constructor
before its body is executed).  This seems to be sufficient at least for
software ISP.

The configuration snippet for logging looks as follows:

  configuration:
    log:
      levels: LIBCAMERA_LOG_LEVELS

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>